+++ src/AssistanceTouristique/ServiceStatSitePOATie.java
@@ -65,7 +65,7 @@ public class ServiceStatSitePOATie extends ServiceStatSitePOA
     /**
      * Operation getStatsSite
      */
-    public AssistanceTouristique.Statistique[] getStatsSite(String date, int idSite)
+    public AssistanceTouristique.Statistique[] getStatsSite(String date, short idSite)
     {
         return _tie.getStatsSite( date,  idSite);
     }
